extract ClusterMarkers into separate component
